requirements.txt
appdirs==1.4.4
argcomplete==1.12.3
autopep8==1.5.7
colorzero==2.0
gpiozero==1.6.2
Pillow==8.3.1
pycodestyle==2.7.0
pyserial==3.5
python-escpos==2.2.0
pyusb==1.2.1
PyYAML==5.4.1
qrcode==7.3
six==1.16.0
toml==0.10.2
